Revert "Fix the 'cvflann::anyimpl::bad_any_cast' error using Lsh" to place it in a different branch
This reverts commit 13efd414c3f811f2bbfb691fd03df530eb492684.

+++ modules/flann/include/opencv2/flann/lsh_index.h
@@ -60,7 +60,7 @@ struct LshIndexParams : public IndexParams
 {
     LshIndexParams(unsigned int table_number = 12, unsigned int key_size = 20, unsigned int multi_probe_level = 2)
     {
-        (*this)["algorithm"] = FLANN_INDEX_LSH;
+        (* this)["algorithm"] = FLANN_INDEX_LSH;
         // The number of hash tables to use
         (*this)["table_number"] = table_number;
         // The length of the key in the hash tables
@@ -94,9 +94,9 @@ public:
     {
         // cv::flann::IndexParams sets integer params as 'int', so it is used with get_param
         // in place of 'unsigned int'
-        table_number_ = get_param<unsigned int>(index_params_,"table_number",12);
-        key_size_ = get_param<unsigned int>(index_params_,"key_size",20);
-        multi_probe_level_ = get_param<unsigned int>(index_params_,"multi_probe_level",2);
+        table_number_ = (unsigned int)get_param<int>(index_params_,"table_number",12);
+        key_size_ = (unsigned int)get_param<int>(index_params_,"key_size",20);
+        multi_probe_level_ = (unsigned int)get_param<int>(index_params_,"multi_probe_level",2);
 
         feature_size_ = (unsigned)dataset_.cols;
         fill_xor_mask(0, key_size_, multi_probe_level_, xor_masks_);
